make the route code cleaner
