+++ src/routes/http-status.ts
@@ -1,0 +1,9 @@
+/**
+ * HTTP status code constants
+ */
+export enum HttpStatus {
+  OK = 200,
+  BAD_REQUEST = 400,
+  NOT_FOUND = 404,
+  INTERNAL_SERVER_ERROR = 500,
+}
