+++ src/routes/index.ts
@@ -1,0 +1,8 @@
+/**
+ * Routes barrel export
+ */
+export { UrlRoutes } from './url.routes.js'
+export { ShortenUrlHandler } from './shorten-url.handler.js'
+export { GetUrlHandler } from './get-url.handler.js'
+export { HttpStatus } from './http-status.js'
+export * from './url-route.schemas.js'
